requirements.txt
aiogram<4.0
yt-dlp>=2025.3.31
PyYAML~=6.0.2